Major reorg of `gempyor.batch`
Split the monolithic `batch` submodule into a subpackage comprised of a
few submodules to assist with review and future development. Addresses
GH-502.
