+++ flepimop/gempyor_pkg/src/gempyor/batch/manifest.py
@@ -1,0 +1,78 @@
+__all__ = ("write_manifest",)
+
+
+import json
+from pathlib import Path
+import sys
+from typing import Any
+
+from ..utils import _git_head
+
+
+def write_manifest(
+    job_name: str,
+    flepi_path: Path,
+    project_path: Path,
+    destination: Path | None = None,
+    **additional_meta: Any,
+) -> Path:
+    """
+    Write job metadata to a manifest file.
+
+    This function produces a manifest metadata file for a batch run. By default the
+    json generated by this function will contain:
+    * 'cmd': The command line arguments provided to the CLI script invoked.
+    * 'job_name': A human readable unique job name.
+    * 'data_sha': The git commit of the project git repository, called 'data' for
+        legacy reasons.
+    * 'flepimop_sha': The git commit of the flepiMoP git repository.
+    Further data can be provided via `**additional_meta`, but these values are
+    overridden by the defaults described above.
+
+    Args:
+        job_name: A user specified or generated from user specified values unique name
+            for the job.
+        flepi_path: The path to the flepiMoP git repository being used.
+        project_path: The path to the project git repository being used.
+        destination: Either a path to where the json file should be written or `None` to
+            write the json file to 'manifest.json' in the current working directory.
+        additional_meta: User specified additional fields added to the manifest json.
+            Values with the name 'cmd', 'job_name', 'data_sha', or 'flepimop_sha' will
+            be overridden by the default behavior. Must be a json encodable type.
+
+    Returns:
+        The path to the written json file.
+
+    Examples:
+        >>> import os
+        >>> from pathlib import Path
+        >>> flepi_path = Path(os.environ["FLEPI_PATH"])
+        >>> project_path = flepi_path / "examples" / "tutorials"
+        >>> manifest = write_manifest("Foobar", flepi_path, project_path)
+        >>> manifest.name
+        'manifest.json'
+        >>> print(manifest.read_text())
+        {
+            "cmd": "",
+            "job_name": "Foobar",
+            "data_sha": "59fe36d13fe34b6c1fb5c92bf8c53b83bd3ba593",
+            "flepimop_sha": "2bdfbc74e69bdd0243ef8340dda238f5504f1ad9"
+        }
+    """
+    flepimop_sha = _git_head(flepi_path)
+    data_sha = _git_head(project_path)
+
+    manifest = {
+        "cmd": " ".join(sys.argv),
+        "job_name": job_name,
+        "data_sha": data_sha,
+        "flepimop_sha": flepimop_sha,
+    }
+    if additional_meta:
+        manifest = {**additional_meta, **manifest}
+
+    destination = Path("manifest.json").absolute() if destination is None else destination
+    with destination.open(mode="w", encoding="utf-8") as f:
+        json.dump(manifest, f, indent=4)
+
+    return destination
