+++ flepimop/gempyor_pkg/tests/batch/test_job_resources_class.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 import pytest
 
-from gempyor.batch import BatchSystem, JobResources, JobSize
+from gempyor.batch import JobResources
 
 
 @pytest.mark.parametrize("nodes", (1, 2, 4, 8))
